refactor(app): slim portal controllers and sync backend findings
