+++ docs/17-backend-annotation-findings.md
@@ -2,13 +2,16 @@
 
 This document records architecture and structure issues that became consistently visible while enriching backend comments. The goal is to preserve concrete observations discovered during code reading, not to propose a full redesign.
 
-## Status Update (2026-03-19)
+## Status Update (2026-03-20)
 
-This document was re-checked after the refactor branch work for findings 1, 2, and 4.
+This document was re-checked after the refactor branch work for findings 1, 2, 3, 4, 8, and 9.
 
 - Finding 1 is now handled in code.
-- Finding 2 is partially handled in code.
+- Finding 2 is substantially improved but still partially handled in code.
+- Finding 3 is improved but still present in code.
 - Finding 4 is now handled in code.
+- Finding 8 is improved but still present in code.
+- Finding 9 is improved but still present in code.
 
 Validation completed on the standard regression path:
 
@@ -19,9 +22,11 @@ Validation completed on the standard regression path:
 Double-check notes:
 
 - The admin-user refactor removed an overlapping, unused application service rather than changing the controller-facing workflow owner.
-- The namespace and skill-lifecycle refactors moved orchestration out of controllers, but preserved the same downstream domain-service calls, request parameters, audit fields, response message keys, and mutation response shapes.
+- The namespace, skill-lifecycle, review, promotion, and compatibility refactors moved orchestration out of controllers, but preserved the same downstream domain-service calls, request parameters, audit fields, response message keys, and mutation response shapes.
 - The security refactor centralized route metadata into one registry, but preserved the same route authorization rules, API-token scope behavior, and CSRF-ignore behavior.
 - `AuthContextFilter` is now scoped to API paths when projecting request attributes. This narrows unnecessary work on non-API requests, but it does not change existing business behavior because `userId` and `userNsRoles` consumers are API-side controllers and interceptors.
+- The localized-exception refactor introduced a shared localized-message contract and moved domain HTTP status ownership into the domain exception types. This reduced handler branching without changing API error codes or HTTP status behavior.
+- The compatibility refactor introduced `ClawHubCompatAppService` and `CompatSkillLookupService` so that repository and visibility-aware lookup logic are no longer duplicated across compatibility controllers and facades.
 
 ## 1. Admin user management is split across overlapping application services
 
@@ -50,7 +55,7 @@ Current state:
 
 ## 2. Several controllers still perform orchestration that belongs in application services
 
-Status: partially handled on branch `docs/backend-annotation-findings-discussion`
+Status: substantially improved but still partially handled on branch `docs/backend-annotation-findings-discussion`
 
 Observed files:
 
@@ -74,10 +79,14 @@ Current state:
 
 - `NamespaceController` has been slimmed down by moving orchestration into `NamespacePortalQueryAppService` and `NamespacePortalCommandAppService`.
 - `SkillLifecycleController` has been slimmed down by moving orchestration into `SkillLifecycleAppService`.
+- `ReviewController` and `PromotionController` have now been slimmed down by moving orchestration into `ReviewPortalAppService` and `PromotionPortalAppService`.
+- `ClawHubCompatController` has now been slimmed down by moving orchestration into `ClawHubCompatAppService`.
 - This branch preserved the original domain-service calls and response contracts for the refactored endpoints.
-- `ReviewController`, `PromotionController`, and `ClawHubCompatController` still exhibit the same structural issue and remain future work.
+- Some controller-side request translation still remains, and other controllers may still mix transport and workflow concerns, so the finding is not fully closed.
 
 ## 3. Compatibility endpoints are tightly coupled to canonical domain and repository internals
+
+Status: improved but still present on branch `docs/backend-annotation-findings-discussion`
 
 Observed files:
 
@@ -93,6 +102,13 @@ Why this stands out:
 Suggested direction:
 
 - Treat compatibility support as a dedicated adapter layer with narrower upstream contracts and fewer direct repository dependencies.
+
+Current state:
+
+- `ClawHubCompatController` no longer coordinates repositories, publish flows, audit logging, and DTO assembly directly. That orchestration now sits in `ClawHubCompatAppService`.
+- A new `CompatSkillLookupService` now centralizes legacy-slug lookup, visibility-aware skill resolution, and latest-version lookup for compatibility use cases.
+- `ClawHubRegistryFacade` now reuses the same compatibility lookup helper instead of duplicating canonical repository access.
+- The compatibility layer still depends on canonical query services, publish services, repository ports, and canonical lifecycle projections. The coupling is narrower and easier to follow, but there is still no fully isolated anti-corruption boundary.
 
 ## 4. Security route policy is spread across configuration and implementation classes
 
@@ -172,6 +188,8 @@ Suggested direction:
 
 ## 8. Exception modeling is duplicated across application, domain, and auth layers
 
+Status: improved but still present on branch `docs/backend-annotation-findings-discussion`
+
 Observed files:
 
 - `server/skillhub-app/src/main/java/com/iflytek/skillhub/exception/LocalizedException.java`
@@ -189,7 +207,16 @@ Suggested direction:
 
 - Keep layer-specific exception types only where they represent a real boundary, and consider converging on a smaller shared contract for localized API-facing errors.
 
+Current state:
+
+- The code now has a shared `LocalizedMessage` contract used across app, domain, and auth exceptions.
+- `GlobalExceptionHandler` now renders localized app, auth, and domain exceptions through one shared rendering path instead of handling each domain subtype separately.
+- Domain localized exceptions now own their HTTP status code, so the handler no longer has to use `instanceof` checks to map domain exceptions to API status codes.
+- Separate exception base types still exist in the app, auth, and domain modules. The duplication is reduced, but the model has not fully converged into one cross-module abstraction.
+
 ## 9. Repository and read-model access patterns are mixed across layers
+
+Status: improved but still present on branch `docs/backend-annotation-findings-discussion`
 
 Observed files:
 
@@ -207,6 +234,13 @@ Why this stands out:
 Suggested direction:
 
 - Define explicit rules for when a use case should depend on domain repository ports, dedicated query repositories, or direct persistence adapters.
+
+Current state:
+
+- This branch reduced some of the most visible mixing at the entrypoint layer by removing direct repository orchestration from `ReviewController`, `PromotionController`, and `ClawHubCompatController`.
+- Compatibility-specific skill lookup and version lookup now live behind `CompatSkillLookupService` instead of being duplicated across compatibility entry points.
+- The broader architectural pattern is still mixed: some flows still use domain repository ports, some use dedicated query repositories, and some app-layer read paths still use direct persistence access.
+- This finding should stay open until the codebase documents or enforces a clearer rule for choosing among those access patterns.
 
 ## 10. OAuth login behavior is decomposed into many small classes without one visible flow owner
