+++ server/skillhub-domain/src/main/java/com/iflytek/skillhub/domain/shared/exception/DomainForbiddenException.java
@@ -8,4 +8,9 @@ public class DomainForbiddenException extends LocalizedDomainException {
     public DomainForbiddenException(String messageCode, Object... messageArgs) {
         super(messageCode, messageArgs);
     }
+
+    @Override
+    public int statusCode() {
+        return 403;
+    }
 }
