+++ server/skillhub-domain/src/main/java/com/iflytek/skillhub/domain/shared/exception/DomainNotFoundException.java
@@ -8,4 +8,9 @@ public class DomainNotFoundException extends LocalizedDomainException {
     public DomainNotFoundException(String messageCode, Object... messageArgs) {
         super(messageCode, messageArgs);
     }
+
+    @Override
+    public int statusCode() {
+        return 404;
+    }
 }
