+++ server/skillhub-domain/src/main/java/com/iflytek/skillhub/domain/shared/exception/LocalizedMessage.java
@@ -1,0 +1,11 @@
+package com.iflytek.skillhub.domain.shared.exception;
+
+/**
+ * Shared contract for exceptions that expose a localized message code and
+ * interpolation arguments regardless of module boundary.
+ */
+public interface LocalizedMessage {
+    String messageCode();
+
+    Object[] messageArgs();
+}
